Supprimer les commentaires et les posts

+++ managePost/deleteComm.php
@@ -1,8 +1,12 @@
-<?php require_once '../MVC/config/connectDatabase.php';
+<?php
+require_once '../MVC/config/connectDatabase.php';
 
     $connexion = connexion();
 
-    $connexion->exec('DELETE FROM croustacomm WHERE id=' . $_GET['commId']);
+    $verifData = $connexion->query('SELECT croustagrameur_id FROM croustapost WHERE id=' . $_GET['commId']);
+    $verif = $verifData->fetch(PDO::FETCH_ASSOC);
+
+    if($verif['croustagrameur_id']===$_SESSION['username']) $connexion->exec('DELETE FROM croustacomm WHERE id=' . $_GET['commId']);
 
     header("Location: pagePost.php?id=" . $_GET['postId']);
     exit();
